Begin connecting with libjpeg

+++ cujpeg/datatypes.h
@@ -13,10 +13,13 @@
 
 typedef unsigned char UINT8;
 typedef signed char INT8;
+
 typedef unsigned short UINT16;
 typedef signed short INT16;
+
 typedef unsigned int UINT32;
-typedef signed int INT32;
+typedef long INT32;
+
 
 #define BLK_LENGTH 64
 #define BLK_LENGTH_BYTE 64*sizeof(INT16)
